Restructure frontend assets and backend code.
