+++ backend/src/implementations/newPost.js
@@ -1,21 +1,19 @@
-import createPost from "../../src/routes/createPost.js";
-import uploadImage from "../../src/routes/uploadImage.js";
+import createPost from "../../src/api/createPost.js";
+import uploadImage from "../../src/api/uploadImage.js";
 
 export default async function newPost(user_id, body, file = null) {
+  if ((user_id, body == null)) {
+    console.log("Null value entered for non-null required parameters.");
+    return;
+  }
 
-    if(user_id, body == null) {
-        console.log("Null value entered for non-null required parameters.");
-        return;
-    }
+  var image_path = null;
 
-    var image_path = null;
+  if (file != null) {
+    image_path = await uploadImage(file);
+  }
 
-    if (file != null) {
-        image_path = await uploadImage(file);
-    }
+  console.log();
 
-    console.log();
-    
-    const resp = await createPost(user_id, body, image_path);
-    
+  const resp = await createPost(user_id, body, image_path);
 }
