+++ docs/frontend.md
@@ -3,9 +3,9 @@
 ## **Technologies Used**
 
 - **Multi-Page Architecture (MPA)** – Separate HTML pages (`index.html`, `profile.html`, `post.html`).
-- **ES Modules (`import/export`)** – Modular JavaScript (`home.js`, `profile.js`, etc.).
+- **ES Modules (`import/export`)** – Modular JavaScript (`index.js`, `profile.js`, etc.).
 - **Fetch API** – Directly communicates with backend endpoints.
-- **Global Styling** – CSS located at `assets/css/styles.css`.
+- **Global Styling** – CSS located at `css/styles.css`.
 
 ## **Implemented Features**
