+++ frontend/css/index.css
@@ -1,6 +1,5 @@
 @import url('https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200');
 
-
 .menu::-webkit-scrollbar {
     display: none;
 }
@@ -118,9 +117,17 @@
 
 /* Base Animations */
 @keyframes fadeIn {
-    0% { opacity: 0; }
-    50% { opacity: 0.3; }
-    100% { opacity: 1; }
+    0% {
+        opacity: 0;
+    }
+
+    50% {
+        opacity: 0.3;
+    }
+
+    100% {
+        opacity: 1;
+    }
 }
 
 @keyframes zoomIn {
@@ -128,10 +135,12 @@
         transform: scale(0.8);
         opacity: 0;
     }
+
     50% {
         transform: scale(0.9);
         opacity: 0.5;
     }
+
     100% {
         transform: scale(1);
         opacity: 1;
@@ -139,23 +148,53 @@
 }
 
 @keyframes colorChange {
-    0% { color: #ff0000; }
-    20% { color: #ff9900; }
-    40% { color: #ffff00; }
-    60% { color: #00ff00; }
-    80% { color: #00ffff; }
-    100% { color: #ff0000; }
+    0% {
+        color: #ff0000;
+    }
+
+    20% {
+        color: #ff9900;
+    }
+
+    40% {
+        color: #ffff00;
+    }
+
+    60% {
+        color: #00ff00;
+    }
+
+    80% {
+        color: #00ffff;
+    }
+
+    100% {
+        color: #ff0000;
+    }
 }
 
 @keyframes fadeInScreen {
-    100% { opacity: 1; }
-    100% { opacity: 1; }
-    100% { opacity: 0.5; }
+    100% {
+        opacity: 1;
+    }
+
+    100% {
+        opacity: 1;
+    }
+
+    100% {
+        opacity: 0.5;
+    }
 }
 
 @keyframes moveBus {
-    from { left: -100px; }
-    to { left: 200%; }
+    from {
+        left: -100px;
+    }
+
+    to {
+        left: 200%;
+    }
 }
 
 /* Base Styles */
@@ -247,13 +286,14 @@ body {
         background: var(--background-dark);
         color: var(--text-dark);
     }
+
     .login-container {
         background: #1f1f1f;
     }
+
     input {
         background: #2d2d2d;
         border-color: #404040;
         color: #fff;
     }
 }
-
