+++ frontend/index.html
@@ -5,8 +5,8 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>AURA</title>
-    <link rel="stylesheet" href="./assets/css/styles.css">
-    <link rel="stylesheet" href="./assets/css/home.css">
+    <link rel="stylesheet" href="./css/styles.css">
+    <link rel="stylesheet" href="./css/index.css">
 </head>
 
 <body>
@@ -14,7 +14,7 @@
         <h1 class="aura-title">AURA</h1>
         <p class="from-text">from</p>
         <p class="magic-bus">Magic School Bus</p>
-        <img src="./assets/img/bus.png" class="bus" alt="Magic School Bus">
+        <img src="./img/bus.png" class="bus" alt="Magic School Bus">
         <div class="wrapper">
             <div class="circle"></div>
             <div class="circle"></div>
@@ -27,7 +27,8 @@
 
     <div class="menu">
         <ul class="menu-content">
-            <li><a href="./pages/login.html"><span class="material-symbols-outlined">home</span><span>Home</span></a></li>
+            <li><a href="./pages/login.html"><span class="material-symbols-outlined">home</span><span>Home</span></a>
+            </li>
             <li><a href="./pages/login.html"><span
                         class="material-symbols-outlined">dashboard</span><span>Search</span></a></li>
             <li><a href="./pages/login.html"><span
@@ -36,7 +37,8 @@
                         class="material-symbols-outlined">analytics</span><span>Notifications</span></a></li>
             <li><a href="./pages/login.html"><span
                         class="material-symbols-outlined">settings</span><span>Create</span></a></li>
-            <li><a href="./pages/login.html"><span class="material-symbols-outlined">person</span><span>Profile</span></a>
+            <li><a href="./pages/login.html"><span
+                        class="material-symbols-outlined">person</span><span>Profile</span></a>
             </li>
             <li><a href="./pages/login.html"><span class="material-symbols-outlined">report</span><span>Shop</span></a>
             </li>
@@ -55,7 +57,7 @@
     </div>
     </div>
 
-    <script type="module" src="./assets/js/home.js"></script>
+    <script type="module" src="./js/index.js"></script>
 </body>
 
 </html>
