+++ frontend/pages/login.html
@@ -1,45 +1,35 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta name="description" content="Login to AURA - Social Media Platform" />
-    <title>AURA - Login</title>
-    <link rel="stylesheet" href="../assets/css/styles.css" />
-    <link rel="stylesheet" href="../assets/css/login.css" />
-  </head>
 
-  <body>
-    <div class="login-container">
-      <h2>Login</h2>
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <meta name="description" content="Login to AURA - Social Media Platform" />
+  <title>AURA - Login</title>
+  <link rel="stylesheet" href="../css/styles.css" />
+  <link rel="stylesheet" href="../css/login.css" />
+</head>
 
-      <form id="login-form">
-        <div class="form-group">
-          <label for="username">Username:</label>
-          <input
-            type="text"
-            id="username"
-            name="username"
-            required
-            placeholder="Enter your username"
-          />
-        </div>
+<body>
+  <div class="login-container">
+    <h2>Login</h2>
 
-        <div class="form-group">
-          <label for="password">Password:</label>
-          <input
-            type="password"
-            id="password"
-            name="password"
-            required
-            placeholder="Enter your password"
-          />
-        </div>
+    <form id="login-form">
+      <div class="form-group">
+        <label for="username">Username:</label>
+        <input type="text" id="username" name="username" required placeholder="Enter your username" />
+      </div>
 
-        <button type="submit">Login</button>
-      </form>
-    </div>
+      <div class="form-group">
+        <label for="password">Password:</label>
+        <input type="password" id="password" name="password" required placeholder="Enter your password" />
+      </div>
 
-    <script type="module" src="../assets/js/login.js"></script>
-  </body>
+      <button type="submit">Login</button>
+    </form>
+  </div>
+
+  <script type="module" src="../js/login.js"></script>
+</body>
+
 </html>
